Hgfs Over VMCI. Asynchronous.
Start supporting Async operations on Hgfs over vmci for Mac.

- Right now when you suspend the VM, change vmx file to disable
vmci - it will correctly fallback to backdoor however it will
not upgrade to vmci when one is available since transport change
is only triggered when you see that current channel is not working.
This logic should go away when session change is implemented which
should detect change in session and restart detection of which channel
is available.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>