Make VSOCK reset connection during connect on QP detach
If the peer of a stream connection detaches from an already
allocated queue pair while the connection state is still
"connecting", we should treat it as a reset. The peer may
detach for a number of reasons, the two primary ones are 1)
the VM got killed, 2) the QP busmem invalidation failed.

This case is already handled by the common code, but not the
linux specific implementation.

This change also resets the VMCI datagram handle allocated by
the stream sockets to VMCI_INVALID_HANDLE when it is
destroyed, since the validity of the handle is used to
determine whether we should call destroy on the handle when
the module is unloaded.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>