Fix Guest Hgfs server interface for multiple callers
Code does not quite cope with multiple callers for the same process
which leads to memory leaks.

The guest channel uses two globals: a channel data object and a server
data object both are shared for all registered connections that use the
internal HGFS RPC. The channel data and server data objects are reference
counted, initialized on the first reference and torndown on the final
reference removal.

At some point in the future if a client requires its own private callback
which will call HGFS server, then the channel code will allocate a new
channel which will in turn get its own private HGFS server session.
Currently this is not supported.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>