Rewrite the ResolutionSet RandR12 code V2
The code was rewritten and added in a new file, resolutionRandR12.c. The rewrite addresses a number of issues:
1) The RandR12 code was never enabled if the server supported the VMwareCtrl extension.
2) If the RandR12 code was trivially enabled by moving the check for the VMwareCtrl extension, vmtoolsd would deadlock and freeze the X server.
This issue is still not fully investigated, though, but commented fixes in this changeset seems to work around it.
3) If the old RandR12 code was enabled together with the deadlock fixes, resolution wasn't set right. In fact mostly it wasn't set at all, perhaps due to races with
other resolutionset mechanisms. Typically the added autofit modes were never used.
4) Error checking and recovery in the old code was not very elaborate.
5) DPI calculation was incorrect since only the pixel dimensions of the first output were considered, and in that case only on resolution increases.
6) The rather long function made bugfixing and error recovery implementation hard. It's now split up.
Together with a minor change in the modesetting kernel these changes appears to remove the need for using VMWareCtrl with the vmwgfx driver, This has only
been tested with two monitors, though.

V2. Fix broken builds when NO_MULTIMON was set to true.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>