lib/file: Reduce log noise
I spotted this in a log file in the PR. It is possible for
one thread to open a lock file on POSIXen and by the time it
tries to stat the lock file another thread has already deleted
(unlinked) the file because it is stale. Had the other thread
not done this, the first aforementioned thread would have done
the unlink.

If this occurs, suppress an warning/error message; it's just
noise.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>