Clipboard gets cleared if pasting a file again
The reason for the bug is that we used to rely on ClipboardClear
signal to check if we are still the clipboard owner. But that signal
is not reliable, and some unexpected ClipboardClear signal may cause
us clearing clipboard data unexpectedly. The fix is to rely on our
own private format like other implementation did.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>