Limit VMCI queue pair memory to 128MB per device
In the future, we will rewrite how the VMCI device allocates queue
pairs, so that they will be located "on the device" and be
exposed through the PCI address range.

We should be able to provide up to a maximum of 128MB queue
pair memory, so this change enforces that limit now.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>