Handle VMCI device shutdown when clients are still attached.
When Windows reboots/halts, it sends the VMCI device
a shutdown PnP event.  This event is sent even if the
device is still referenced.  Clients are supposed to
handle IRP_MJ_SHUTDOWN (or similar) and detach, but if
they don't, then it's possible for them to call us
after the device is gone, at which point we can
explode.

Fixed so that we gracefully handle this.  On a shutdown
event:

1) Fail VMCI handle creation from this point on, but
   allow other calls, so that handles can be closed.
2) Inform vsock that the device is going away.
3) VSock fails DGRAM/STREAM creation from this point on.
4) VSock informs all open sockets that the device
   is going away.  The sockets close their handles
   and their wait list is kicked so that they wakeup
   from recv() etc.
5) VSock closes the control channel.
6) VSock fails all creation calls from this point on,
   but allows other calls, so that sockets can be
   closed etc.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>