lib/lock: simplify rank checking while panicing
If we detect a rank violation in the locking code and we're not
already panicing, making a list of the locks held is quite
appropriate; the locking code will then explcitly panic.

If we panic due to a non-locking cause, rank violation checking and
list printing (but not panicing) can get pretty noisy - we have
ranks screwed up as we try to panic and/or coredump.

If the locking code is the source of a panic due to a rank violation
only print the first list; other things are highly likely to be
noise.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>