Panic_Panic needs to set mxInPanic flag
The userlock module relies on this flag to not get into a panic loop
from lock rank violation when we are trying to dump core.  This is
the default Panic() implementation used by UI and other apps.  For
other apps that stub out Panic(), they are not likely to use MXUser
locks.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>