Always register VMCI Socket control message handlers on Linux
The VMCI Socket code on Linux only registers an address family
with the kernel as long as there are active sockets. As part
of registering/unregistrering the address family, the VMCI
Socket module also registers/unregisters the stream protocol
control message handlers. This results in no control message
processing being done, when there are no active sockets. For
peers attempting to connect this results in control messages
being silently dropped, and in the case of a connect, the
connect will just hang. This change moves the registration of
the control messages to the vsock module initialization, such
that control messages will always be handled even if there are
no active sockets.

Note that this change also makes the VMCI Socket module
loading fail, if the VMCI device isn't available. This isn't
strictly necessary, but the vsock module isn't all that useful
without the VMCI device.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>