Reference count the VMCI device.
We must reference the device when we have clients
attached, for example HGFS or vsock.  Changed
VMCI_DeviceGet() and VMCI_DeviceRelease() so that
they reference count, and added new function
VMCI_DeviceInUse() so that we can tell when the
device is referenced during a query-for-stop IRP
and fail if necessary.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>