Hang in VMCISockets caused by uninitialized notification state.
It turns out that a recent unload/shutdown fix broke the
notification code.

We don't call the notification functions when the qpair
is invalid, to guard against touching the qpair after
the device has shutdown, but some of those functions
need to be called *before* a qpair is actually attached.
There are actually a few of them, so rather than introduce
a new macro for notifications that don't need qpairs,
I've removed the qpair check from the macros and embedded
it as necessary in the notification code.

We also uncovered a problem where if the notification
write window is less than a page we can underflow.
Fixed on all platforms.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>