Implement capabilities for DnD/CP.
Implement capabilities exchange and add support for new isolation
settings in VMX. The general idea is that each side (host, guest)
transmits to the controller its capabilities (DnD/CP support,
ability to provide certain clipboard types) in a PING request. VMX
file settings can be used to disable a variety of clipboard content,
direction combinations. For example, RTF DnD in the guest to host
direction can be disabled. In the PING response, controller will
provide a requisition to the request, telling the host or guest
what types of DnD/CP content are allowed on the clipboard. When
preparing the clipboard for transmission, the host or guest checks
this requisition and will not supply formats that are not allowed.

The controller will enforce this by removing any clipboard contents
that do not match the published requisition, so in case the guest
or host does not do the right thing, or ignores the requisition in
the ping reply, the controller will catch it. If any of this
processing causes the clipboard received by the controller to be
empty, the DnD or CP operation does not continue.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>