Fix closing of open handles for searches on "Shared Folders".
Our HGFS server callback from the vmdb for UI updates always
force closes any open handles that are not on any shares.

This was done even if no shares had been added or removed.

This had the side-effect of alwasy closing the handles of
searches for the base of the name space i.e. "Shared Folders"
folder. This meant a race where those clients could end up
getting invalid handle returned for a directory read of the
virtual folder. It is better to return possibly stale (unlikely)
shares for that directory read than always fail even if
no shares changed.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>