Fix KDE file copy and paste.
There were a couple of bugs in how we respond to paste requests.
First, we can be called multiple times. So we needed a guard to
check if we are already in the process of transferring files.

Second, we should never clear the selection data unless we see an
error or fail case of some kind. While it does not affect GNOME
environment, looks like KDE can get really confused by this (the
symptom is a prompt for path name which has been reported often
in bugs over the past year I have been involved with DnD.)

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>