Allow VMCI Socket connections between VM and host app of same user
VSockets created by a host process should be trusted to talk
to a VM started by the same user. This change implements this
by:

1) Storing a user identification of the user that created the
socket (the owner).
2) Making child socket inheret the owner.
3) Modify the notion of trusted in the VMCI Socket world to
include sockets created by the same user as the owner of
the VM.

Concerning step 3: Since the VMCI Socket modules notion of
trusted is now more inclusive than that of the VMCI kernel
module, all datagram handlers on the host are now registered
as trusted by the VMCI Socket module. The datagram handlers
themselves now perform the trusted check, and discards
datagrams accordingly. This was already the case for the
control datagrams for VMCI sockets, but now regular VMCI
Datagram sockets do the same. Since the check is not free (the
VMCI kernel module is likely to access hash tables and grab
locks to determine the owner of a context) - a one entry
"trust" cache has been created for each socket.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>