Unify the VMCI datagram code.
We now use the host datagram code, with resources and
contexts, for both drivers, with only a few tweaks.
We have a new component that performs routing, and using
the decision from that, we dispatch to host or guest as
appropriate.

There are obviously still some TODOs.  We don't yet
create a context for the guest itself when acting as
a guest, and this is necessary to get the privs flags
right; for now, we steal the original guest code for
that.  And getting the context ID is still determined
by which driver we build.

Most of this change is in vmciDatagram.c and vmciRoute.c;
the rest was necessary to get the files to compile for
both drivers, on all platforms.  Bleh.

Also cleaned up a bunch of whitespace/tabs.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>