Make Linux reference count vsock module on file ops.
Bug 681873 reports that the linux vsock module can be removed
while the vsock device is opened. The vsock file operations
registered with the device do not specify the module itself as
the owner, and therefore the kernel file operation code cannot
correctly reference count the module on open/close. This
change adds the vsock module as owner, which fixes the
issue. All other function tables registered, e.g., the socket
operations, do set the owner field (as does the VMCI kernel
module).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>