Register new HgfsSeek function in file_operations structure.
Please check bug 667029 for more details. On Ubuntu 2.6.37, seek()
functions for any file residing in HGFS share fail with an error.
We are not registering any explicit function for llseek operation in
file_operations structure. This could be an issue with Ubuntu latest
release. Modified the code to design a new HgfsSeek function [ a
thin wrapper over generic_file_llseek()] and register that for
llseek operation in file_operations.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>