Unify VMCI guest and host queue pair code.
This change unifies the two VMCI queue pair implementations
for guest and host. The main part of the change is moving the
guest functions to the host module, and add the use of the
VMCI routing function in the vmciQPair module to correctly
select either guest or host routes. For vmkernel, we still
disable most of the guest functionality, since the vmkernel
has its own implementation of much of the queue pair code
anyway. To reduce the vmkernel only part of the queue pair
code is left as a future clean up task.

Other notable changes were that for both Linux and Mac, the
host and guest tracked queue pairs differently
(VMCIQueueKernelIf). For Linux, it was straight forward to
combine the two at the cost of adding a pointer to the
VMCIQueueKernelIf structure in the guest. For Mac OS, the host
and the guest take completely different approaches, so the
unified VMCIQueueKernelIf is now a union of the host and guest
versions.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>