Introduce a connection timeout for VMCI stream sockets
When attempting to connect a VMCI Stream socket to a context,
where VMCI Sockets isn't loaded, the connecting process will
hang (or time out if the process itself has specified a
timeout on the connect). This change makes all connection
attempts default to time out after 2 seconds (on ESX, we
should be able to do about a thousand connection attempts in
that time period). For asynchronous connects, we schedule a
timeout function, that changes the socket state on failed
connection attempts and signals the error to the owner.

The connect timeout can be changed through a new VMCI Socket
specific socket option SO_VMCI_CONNECT_TIMEOUT. Testing of
this socket option and the timeout in general has added to the
sockoptVmciTest in the vsock test suite.

This change also makes VMCI Sockets on vmkernel report socket
errors in the read set on select to make UWs compatible with
Linux.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>