Unity/X11: Handle window titles encoded with COMPOUND_TEXT encoding.
COMPOUND_TEXT is a string type used to encode characters outside the
ASCII range, and it may be used to encode window titles.  Unity/X11
ignored all such titles, so the host UI assumed there was no title.
Whoops.

This change employs GDK to do the heavy lifting for us.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>