Remove vmciGuestKernelIf.c on Linux, Windows, Solaris
This change moves the functionality of vmciGuestKernelIf.c to
vmciKernelIf.c on all guest platforms except MacOS. On MacOS,
the functionality in vmciGuestKernelIf.cpp is more tightly
coupled with the methods exported by the driver object, so
that is left alone for now. Added compat_pci.h and
compat_ioport.h to host driver tarball.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>