Translate EEXIST into ENOTEMPTY for Posix guests.
When the user executes DeleteDirectoryInGuest api with recursiveflag set to
FALSE, vix code calls File_DeleteEmptDirectory() and it internally calls
rmdir() function. If the specified directory is not empty, rmdir fails and
errno is set to different values on linux and solaris. To maintain
consistency, modified the code to translate EEXIST into ENOTEMPTY error
for Posix guests.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>