GHI/X11: Scale icons to fit in GuestMsg transport.
GHI/X11 used to scale icons down to fit over the backdoor.  The scaling
was stripped when alternate icon transport was introduced for Windows,
since transport details shouldn't be handled at this level.

In a perfect world, I'd leave it to GHI's TCLO layer to handle scaling
to fit within GuestMsg limits, but that's more of a PITA to add than
it's worth.  So, instead, I'm taking the easy/lame way out and down-
scaling icons at the source, relying on GdkPixbuf's scaling routines
to do the heavy lifting.  (NB:  This is just bringing back our old
scaling code.)

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>