This is part 14 of a change to support Nested VMs.
This part adds more support for the host device in the
guest driver.  You can now add a host device to a guest.
The VMX IOCTLs will still fail, but at least we do
the access checking now.

This also adds all the new device names, along with the
general interface, which allows things like getting the
CID, attaching to vsock etc.

One final change is to the wrapper for allocating
kernel memory.  See the comment in vmci_kernel_if for
a full explanation.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>