Extend vmciRoute with hypervisor to host routing.
The vmci routing function was defaulting to
VMCI_ROUTE_AS_GUEST, when the hypervisor is the source and the
host is the destination. However, vmci events generated
locally on the host are sent from hypervisor to host, and
should be routed as host communication. Also, a guest would
never use the hypervisor context ID as the source, so there is
no ambiguity.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>