Merge Linux VMCI guest and host driver code
This change merges the code for the Linux specific part of the
VMCI driver. The change has two parts: one is the merging of
the actual code of the drivers, and the other is changing the
host and guest makefiles to pick up the VMCI driver files from
the same place.

The actual merging of the Linux driver code is mainly a code
movement from the guest driver to the host driver. Currently,
the driver will still be either guest or host, based on the
flag isGuestDriver which gets controlled by whether
VMX86_TOOLS is defined. All code is now compiled regardless of
whether the driver is guest or host.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>