Fix a crash when calling RpcIn_stop.
If an RpcInLoop callback calls RpcIn_stop, then the state of the RpcIn
struct becomes corrupt (we have a timer but no channel), leading to a
crash the next time the RpcInLoop runs. One way this can happen on
Windows VMs is if we pump the main thread's message queue from inside
an RPC handler, since we will call RpcIn_stop in response to a
WM_WTSSESSION_CHANGE message.

To fix this, we set an inLoop var in the RpcIn struct at the start of
the RpcInLoop function. RpcIn_stop checks the var, and if it is true,
it sets a shouldStop to TRUE instead of stopping the channel itself.
When we eventually exit the RpcInLoop function, we check shouldStop
and stop the RPC channel if it is true.

This obviously only works if RpcIn_stop and RpcInLoop are called on
the same thread, but RpcIn is not thread-safe to being with, so there
didn't seem to be any point in trying to fix this problem for that
case.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>