Make VMCI Stream Sockets use control packet destination for binding.
When creating a VMCI stream socket connection, the protocol
calls VMCI_GetContextID() to set the local CID of a
connection. However, in the nested VM case, this may no longer
be correct, since a given connection can either be a host CID
endpoint or a guest CID endpoint. Instead, we can use an
incoming control packet to determine what destination CID the
peer is using when sending to us, and use that as our local CID.

For server side sockets, the change is simple - instead of
calling VMCI_GetContextID() to set the local CID when
receiving a request packet, we use the dst cid of the
packet. For the client side, we delay the assignment of the
local CID until we receive the negotiate packet from the
server side. Until then, a connecting socket will continue to
have the local CID of any - just like a listening
socket. Since all sockets share the same port range regardless
of their CID, no additional port conflicts will arise from
this.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>