unity/X11: Exit Unity upon notice of session termination.
During session termination, not exiting Unity until vmusr unloads the
plugin is racy.  By the time the Unity shutdown routine executes, it's
possible that the X server has already terminated (or at least severed
our connection to it).  Instead of waiting until then, we'll instead
exit Unity as soon as vmusr is alerted of a pending session termination,
_before_ vmusr returns its acknowledgement to the session manager.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>