desktopEvents: Tell the session manager it should not start us when relaunching a session.
A session as defined by an X Session Manager is a collection of applications and
some state such that this collection may be recreated arbitrarily in the future.
For example, a developer may define a development session consisting of an
instance of Emacs and a web browser pointing to some documentation.  Rather than
relaunch both apps upon each login, s/he could define a "happyfundev" session
and select that to launch that collection of apps at any time.

One requirement of being managed by a session manager is that one must provide
add'l metadata to the session manager to support this restart/restore case.
However, when the session manager relaunches applications as part of a recorded
session, it doesn't record anything about the environment -- instead, each
application must persist as much of the environment it needs by itself.

In our case, we rely on the environment set up by (ironically) the session
manager during its autostart phase in order to launch correctly.  (E.g., we
rely on environment variables like KDE_FULL_SESSION to determine that we're
running under KDE.)  Since vmusr gains nothing from being restartable via the
session manager, we'd rather inform the session manager NOT to restore vmusr
as part of any desktop session.  Instead, it should continue to only launch
as part of the XDG autostart phase.

Without doing this, there is a race between two instances of vmusr launching:
one from the autostart phase and one from the session restore phase.  If the
session restore instance wins, things break.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>