Make VMCI Linux driver support host and guest concurrently
This change makes the VMCI Linux driver enable both the guest
and host driver functionality as appropriate. The host driver
part will always be initialized, but to limit its exposure in
guests not used as a virtual host as well, we adopt the
following behaviour for VMCI_HasHostDevice:
- if there is no VMCI PCI device, VMCI_HasHostDevice()
returns TRUE if it has been initialized.
- if there is a VMCI PCI device, VMCI_HasHostDevice()
returns TRUE iff theere are active contexts.
This will make the default behaviour in non-nested cases the
same as prior to this change.

In the nested VM case, we may have VMCI clients, that use
resources as a host endpoint, before any contexts have been
created, e.g., allocating a doorbell before any VMs are
started. The VMCI_HasHostDevice() check has been removed from
the doorbell creation code to allow for this case.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>