This is part 17 (this is getting ridiculous) of a change to support Nested VMs.
This change switches over the HasPeronality() and
DeviceEnabled() rules to match those of the Linux
driver.  This allows it to handle *always* having
two devices in a guest, even if a hosted product
is not installed, which is now a restriction
imposed by the installer.

DeviceShutdown() and InUse() are redundant on
platforms that don't have retarded stop/removal
rules, so moved that functionality out of the
common code and	into the Windows driver code.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>