Fix non-glib rpcin code.
For non-glib tools, set in->nextEvent to NULL as soon as we enter
RpcInLoop. This is important, since in->nextEvent and the next
pointer in EventManager_ProcessNext point to the same Event struct
when we enter the function, and EventManager_ProcessNext frees the
pointer after RpcInLoop exits, so freeing it in RpcInLoop results in
a double-free and corrupts the heap. This can happen if RpcInStop is
called at the end of the loop in response to a call to RpcIn_stop
from inside cb->callback.

Also, after re-reading the glib docs and consulting with others, fix
how we manage the lifetime of the GSource used to run the RpcInLoop
in the glib case. If we are stopping the loop from _inside_ the
loop, then we need to unref the source but _not_ destroy it; instead
we should return FALSE so that the g_main_loop with also unref it
(and hence destroy it). If we are stopping the loop from outside the
loop, then the source must be destroyed (to get the g_main_loop to
release its ref) and then unref'd (to release our ref). Finally if
we do not reschedule the timer while running the loop (because the
timeout hasn't changed) then we must return TRUE so that the source
will remain attached to the g_main_loop.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>