Allow shrinking ZFS partitions.
Also reorganize the code a little bit. ZFS doesn't really look like a
normal disk, so we'll just assume it's living in a VMDK (not that we
really check for that for other FSes). Users can probably shoot
themselves in the foot a lot easier with ZFS by attaching storage
to existing volumes and doing other crazy things ZFS allows, but
hey, if they want to do it...

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>