VMCIQPair_Detach() should not leak when it fails.
We're hitting an ASSERT on shutdown. Code can
still have a socket open very late in the OS
shutdown sequence.  VMCISockets forcibly closes
this socket so that it can safely unload.  The
device I/O space is already gone at this point,
and with it the queuepairs.  So we fail to
detach the queuepair, but we still ASSERT at
the VMCISocket level that we released it.
This occurs in both the common code (ESX,
Mac OS and Windows) and the Linux code.

At the VMCI level, we're actually leaking
when it fails, because we only release the
entry and qpair struct on success.  There's
nothing a caller can do at that point, so we
should always release the entry and qpair
struct.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>