fix irq naming to match network convention
The vmxnet driver always requests irq with name of 'vmxnet' which is
incorrect. It causes multiple entries in /proc/interrupts and confuses
scripts like irqbalance that try and manage smp affinity on network
cards.

The fix is simple, just use the standard convention of putting device
name in the irq name. This applies to all kernel versions.

Signed-off-by: Stephen Hemminger <shemminger@vyatta.com>
Signed-off-by: Shreyas N Bhatewara <sbhatewara@vmware.com>
Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>