Add API to query VMCI sockets driver version.
VMCF asked for a way to determine the version of
the VMCISockets driver so that they can blacklist versions
that have issues.  This change introduces such an API.

The version is platform-dependent, like the VMCI
driver version, and uses the same format, A.B.C.D,
where A is an epoch, B the major version (incremented
per release branch), C the minor version (incremented
per change) and D is unused.  It is embedded in the
file vsock_version.h for each platform.

The API is exposed through a new call in vmci_sockets.h,
VMCISock_Version(), as an unsigned integer.  New macros
are provided to extract the A, B and C component parts
of the version.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>