Enable logging by default in vmtoolsd.
To ensure we get at least some logging by default, enable a default logger
that writes to syslog in vmtoolsd. The default log level ("message" on beta
builds, "warning" on release builds) is unchanged.

The change also makes "syslog" the default log handler for everybody,
in case a handler is not specified in the config data. (Previously, it
would be "stdout" on POSIX, or "outputdebugstring" on Win32.)

Also:
. reset the logging system after reading the command line args, so that
syslog can pick up the change in the container name.
. fix an issue in the POSIX syslogger; openlog() doesn't seem to make its
own copy of the "ident" argument, so it could end up referencing freed
memory as we reconfigure the log system. Keep a copy in the log data.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>