Reduce memcpy for data packets on VMCI requests
The get data packet routine would always do a copy of the guest
data into the memory allocated buffer even if the request was
for a writable memory buffer. So for V4 read and directory read
requests there was an unnecessary memcpy.

The reason for this was that the meta packet uses the same routine
but does need to read in the guest memory (for the request) but
needs to write the reply into the same memory. So instead I have
now added a third buffer access option to the READ and WRITE existing
ones: READ-WRITE. So now meta packets use the READ-WRITE and the
V4 read and directory read will use the WRITE option. The memcpy
to read the guest data is only performed for READ or READ-WRITE
buffer requests.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>