Fix ordering of HGFS opcode for search read V4 backwards compatibiliy
Since change notification is released as of Beta 1 the previous change
of inserting the search read opcode ahead of this will lead to
incompatibily for Beta 2 servers against Beta 1 tools.

Make the continuity seamless by keeping the change notification
opcode ahead of the search read V4. This means the HGFS server dispatch
handler must check for a NULL entry in its handler dispatch table now.
(Or we could put a dummy dispatch handler which always returns protocol
error, but I think NULL is okay.)
The change notification opcode is only for requests sent by the server
to the guest clients in response to a directory change event.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>