On running queuepairAllocTest24_l test on Ubuntu9 guest, the guest goes unresponsive
We try to allocate the queuepair pages before we make the hypercall (which would
eventually hit the restriction in the device).  On Windows and Mac OS, we can't
allocate a single contiguous block this big, so the allocation fails immediately.
On Linux, we allocate each page separately, so what happens is that the system
starts thrashing and becomes unresponsive.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>