Call g_clear_error(&err) after g_locale_from_utf8() fails.
After catching an error from g_locale_from_utf8() the error structure
is not freed. This results in a warning from GLib when we later call
g_spawn_async() with the same (now populated) error parameter.

While we're in here copy the debug message from the Windows version of
PowerOpsScriptCallback() so that the exit code of the power script is
logged.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>