lib/file: prep change before overhaul
The "make temp" functions currently live in file.c and should
live in fileTemp.c. In order to move them, remove the Msg_*
stuff - this should be Log and Warning - and include the
work-around just added to ESX50. After this change and moving
the code to its proper place, a full fix for the worked-around
problem will be done.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>