Fix VMCI doorbell index allocation
With the merging of the guest and host code, the code that
allocates a doorbell got shuffled around. This introduced a
bug, where a guest personality allocation of a doorbell will
register the index of a doorbell handle with the VMCI backend
before allocating it. This change moves the allocation of the
index up before it is registered with the backend. To avoid
generating callbacks before we know for sure that a given
doorbell has been successfully created (this can happen due to
index sharing), an active flag has been added to the doorbell
entry. This flag is set to true only after the creation of the
doorbell has been successfully completed.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>