VMCI device needs a limit on the datagram queue
It turns out that we've never put a limit on the number of
datagrams that can be queued in the device.  The perf test
keeps pushing datagrams, and with nobody reading them, the
queue in the VMX just keeps growing until we exhaust the
pagefile.

Added a limit in VMCI_QueueDatagram().  We use the same
limit we use elsewhere, which is 2 x DG_MAX, or ~128K, for
normal datagrams, and a higher limit for events (datagrams
sent from the VMX/event resource ID).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>