Fix saving MS Project files in a folder on a NFS server
The files appeared as read only because the set attributes
call to update the file modification time was failing.
This occurred due to the Posix HGFS server switching to
running as super user when updating the file timestamps by
calling futimes. This was done if the VMX was the file owner
or running as root. This is okay for local file systems but
not for network shares. In this case the VMX was indeed the
file owner and so switching to super user was not necessary.

Fix is to not switch to super user if the VMX is the file
owner.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>