GHI/X11: Don't pass DESKTOP_AUTOSTART_ID env var to children.
DESKTOP_AUTOSTART_ID was proposed on the xdg@freedesktop.org mailing list,
but doesn't seem like it made it to a final spec.

http://lists.freedesktop.org/archives/xdg/2007-January/007436.html

It refers to a XSMP session manager client ID which shouldn't be
passed to children.  Having this environment variable breaks launching
nautilus w/o arguments.  (Aside, GNOME fixed this upstream in response to
https://bugzilla.gnome.org/show_bug.cgi?id=649063.)

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>